Remove screen.kt and use only renderScreen.kt
Use only one composeable to achieve the same, not need
for the wrapper.
